added environment compensation for co2 sensor

+++ main/common.h
@@ -9,6 +9,6 @@
 #include "freertos/FreeRTOS.h"
 #include "freertos/semphr.h"
 
-int get_time_micros(uint32_t t);
+uint64_t get_time_micros(uint32_t t);
 
 #endif
