+++ main/main.c
@@ -12,7 +12,7 @@
 #include "ccs811_sensor/ccs811.h"
 
 
-int get_time_micros(uint32_t t) {
+uint64_t get_time_micros(uint32_t t) {
     return 1000000 * t;
 }
